+++ minishell/Makefile
@@ -11,6 +11,7 @@ SRC := src/main/main.c \
 		src/main/main_utils.c \
 		src/main/main_utils1.c \
 		src/executor/executor.c \
+		src/executor/executor_utils.c \
 		src/executor/executor_utils1.c \
 		src/executor/executor_utils2.c \
 		src/executor/executor_utils3.c \
